refactor: refactor formatter functions and sort imports
