+++ src/lib/dataFormatters.ts
@@ -1,14 +1,16 @@
 import { DateTime } from 'luxon'
 import { PhoneNumber } from '@/common/types'
 
-export const formatDateTime = (date: string | null): string => {
+export const formatDateTime = (date: string | null | undefined): string => {
   if (!date) return ''
   return DateTime.fromISO(date, { zone: 'UTC' })
     .setZone(DateTime.local().zoneName)
     .toFormat('dd.MM.yyyy HH:mm')
 }
 
-export const formatPhoneNumber = (phoneNumber: PhoneNumber | null): string => {
+export const formatPhoneNumber = (
+  phoneNumber: PhoneNumber | null | undefined
+): string => {
   if (!phoneNumber) return ''
   const { countryCode = '', lineNumber } = phoneNumber
   const fullNumber = countryCode + lineNumber
@@ -18,7 +20,9 @@ export const formatPhoneNumber = (phoneNumber: PhoneNumber | null): string => {
   )
 }
 
-export const formatReferenceNumber = (number: string | null): string => {
+export const formatReferenceNumber = (
+  number: string | null | undefined
+): string => {
   if (!number) return ''
   return number.replace(/(\d)(?=(\d{3})+$)/g, '$1 ')
 }
