+++ src/modules/roles/components/columns.tsx
@@ -1,12 +1,12 @@
+import { Sort } from '@/common/types'
+import DataTableSort from '@/components/ui/data-table-sort'
+import Status from '@/components/ui/status'
+import { fallbackStatus } from '@/constants/fallBackStatus'
+import { formatDateTime } from '@/lib/dataFormatters'
+import { getSortState } from '@/lib/getSortState'
 import { Column, ColumnDef } from '@tanstack/table-core'
 import i18next from 'i18next'
-import { formatDateTime } from '@/lib/formatDateTime'
-import Status from '@/components/ui/status'
-import DataTableSort from '@/components/ui/data-table-sort'
-import { Sort } from '@/common/types'
 import { roleStatuses } from '../constants/statuses'
-import { fallbackStatus } from '@/constants/fallBackStatus'
-import { getSortState } from '@/lib/getSortState'
 
 export interface Role {
   id: string
